+++ src/Components/Commons/Layouts/index.tsx
@@ -1,5 +1,6 @@
 import { Footer } from '../footer/footer';
-import { Header } from '../header/header';
+import { Header } from '../header';
+
 
 interface TProps {
   children: JSX.Element;
