+++ src/Containers/main/main.tsx
@@ -1,7 +1,7 @@
-import { CityTabs } from '../../components/main/sity-tabs/sity-tabs.tsx';
-import { PlacesSorting } from '../../components/main/places-sorting/places-sorting.tsx';
-import { PlaceCard } from '../../components/main/place-card/place-card.tsx';
-import { CityMap } from '../../components/main/city-map/city-map.tsx';
+import { CityTabs } from '../../components/main/sity-tabs/index.ts';
+import { PlacesSorting } from '../../components/main/places-sorting/index.ts';
+import { PlaceCard } from '../../components/main/place-card/index.ts';
+import { CityMap } from '../../components/main/city-map/index.ts';
 
 import { IPlaceCardProps } from '../../types/index.ts';
 import placesData from '../../assets/data.json';
